backend/pyproject.toml
[project]
name = "tote-inventory-backend"
version = "0.1.0"
description = "FastAPI backend for Tote Inventory"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "SQLAlchemy==2.0.34",
    "pydantic==2.9.2",
    "python-multipart==0.0.12",
    "pillow==10.4.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "email-validator==2.2.0",
]

[tool.setuptools]
packages = ["app"]

[tool.uv]
index-url = "https://pypi.org/simple"
